Correct parsing with century (C) and 2-digit year (YY) directives
Year must be computed using century numbering starting with zero.
`Timex.century/1` returns the century numbered in the standard way
(starting from the 1st century, with XX01 being the first year in
the century), so it messes up the calculation, and can't be used.

Also to be in line with how `Timex.format/2` handles the `C`
directive as well as glibc `strptime(3)`, centuries should be
numbered from zero.

For the 2-digit year (`YY`) the fix is very minor since the
parsing depends on current time, and the first time the fix makes
a differnce is in year 2100, in which parsing `"00-01-01"` with
the format `"{0YY}-{0MM}-{0DD}"` will yield 2100-01-01, which
is probably what is desired instead of 2000-01-01.